Factor out FilterToolbar
