+++ src/app/plan/components/Wizard/VolumesTable.tsx
@@ -18,22 +18,7 @@ import {
   SelectOptionObject,
   Level,
   LevelItem,
-  DataToolbar,
-  DataToolbarContent,
-  DataToolbarToggleGroup,
-  DataToolbarItem,
-  DataToolbarFilter,
-  Dropdown,
-  DropdownToggle,
-  DropdownItem,
-  InputGroup,
-  Select,
-  SelectOption,
-  SelectOptionProps,
-  TextInput,
-  ButtonVariant,
 } from '@patternfly/react-core';
-import { FilterIcon, SearchIcon } from '@patternfly/react-icons';
 import { Table, TableVariant, TableHeader, TableBody } from '@patternfly/react-table';
 import ReactJson from 'react-json-view';
 import { WarningTriangleIcon } from '@patternfly/react-icons';
@@ -41,6 +26,10 @@ import spacing from '@patternfly/react-styles/css/utilities/Spacing/spacing';
 import SimpleSelect from '../../../common/components/SimpleSelect';
 import { usePaginationState } from '../../../common/duck/hooks';
 import { IFormValues, IOtherProps } from './WizardContainer';
+import FilterToolbar, {
+  IFilterCategory,
+  FilterType,
+} from '../../../common/components/FilterToolbar';
 
 const styles = require('./VolumesTable.module');
 
@@ -140,24 +129,6 @@ const VolumesTable: React.FunctionComponent<IVolumesTableProps> = ({
 
   ////////////////////////////////////////////////////////////////////////////////
 
-  enum FilterType {
-    select = 'select',
-    search = 'search',
-  }
-
-  interface IFilterCategory {
-    key: string;
-    title: string;
-    type: FilterType;
-    selectOptions?: SelectOptionProps[]; // TODO only if select type?
-    placeholderText?: string; // TODO only if select type?
-  }
-
-  interface IFilterControlProps {
-    category: IFilterCategory;
-    value: any; // TODO type this... SelectProps.selections? string? array?
-  }
-
   const filterCategories: IFilterCategory[] = [
     {
       key: 'one',
@@ -188,129 +159,13 @@ const VolumesTable: React.FunctionComponent<IVolumesTableProps> = ({
       placeholderText: 'Filter by three...',
     },
   ];
-  const [currentCategoryKey, setCurrentCategoryKey] = useState(filterCategories[0].key);
-  const [isCategoryDropdownOpen, setIsCategoryDropdownOpen] = useState(false);
+
   const [filterValues, setFilterValues] = useState([]);
-
-  const onCategorySelect = category => {
-    setCurrentCategoryKey(category.key);
-    setIsCategoryDropdownOpen(false);
-  };
-
-  const clearAllFilters = () => console.log('clearAllFilters!'); // TODO
-  const onClearFilter = () => console.log('onClearFilter'); // TODO
-  const onFilterChange = () => console.log('onFilterChange'); // TODO
-
-  const currentFilterCategory = filterCategories.find(
-    category => category.key === currentCategoryKey
-  );
-
-  const FilterControl: React.FunctionComponent<IFilterControlProps> = ({ category, value }) => {
-    const [isFilterDropdownOpen, setIsFilterDropdownOpen] = useState(false);
-    // @ts-ignore
-    const onFilterSelect = () => console.log('onFilterSelect', arguments);
-
-    // TODO split into sub components?
-
-    const [inputValue, setInputValue] = useState('');
-
-    const onInputSubmit = () => console.log('TODO');
-
-    // @ts-ignore
-    const onInputChange = () => console.log('onInputChange', arguments);
-
-    const onInputKeyDown = () => {
-      onInputSubmit();
-      // @ts-ignore
-      console.log('onInputKeyDown', arguments);
-    };
-
-    if (category.type === FilterType.select) {
-      return (
-        <Select
-          aria-label={category.title}
-          onToggle={() => setIsFilterDropdownOpen(!isFilterDropdownOpen)}
-          onSelect={onFilterSelect} // TODO ???
-          selections={value}
-          isExpanded={isFilterDropdownOpen}
-          placeholderText="Any"
-        >
-          {category.selectOptions.map(optionProps => (
-            <SelectOption {...optionProps} />
-          ))}
-        </Select>
-      );
-    }
-    if (category.type === FilterType.search) {
-      const id = `${category.key}-input`;
-      return (
-        <InputGroup>
-          {/*
-            // @ts-ignore issue: https://github.com/konveyor/mig-ui/issues/747 */}
-          <TextInput
-            name={id}
-            id={id}
-            type="search"
-            aria-label={`${category.title} filter`}
-            onChange={onInputChange}
-            value={inputValue}
-            placeholder={category.placeholderText}
-            onKeyDown={onInputKeyDown}
-          />
-          <Button
-            variant={ButtonVariant.control}
-            aria-label="search button for search input"
-            onClick={onInputSubmit}
-          >
-            <SearchIcon />
-          </Button>
-        </InputGroup>
-      );
-    }
-    return null;
-  };
-
-  const filterToolbar = (
-    <DataToolbar id="pv-table-filter-toolbar" clearAllFilters={clearAllFilters}>
-      <DataToolbarContent>
-        <DataToolbarToggleGroup variant="filter-group" toggleIcon={<FilterIcon />} breakpoint="xl">
-          <DataToolbarItem>
-            <Dropdown
-              toggle={
-                <DropdownToggle onToggle={() => setIsCategoryDropdownOpen(!isCategoryDropdownOpen)}>
-                  <FilterIcon /> {currentFilterCategory.title}
-                </DropdownToggle>
-              }
-              isOpen={isCategoryDropdownOpen}
-              dropdownItems={filterCategories.map(category => (
-                // TODO properties for row property name and title
-                <DropdownItem key={category.key} onClick={() => onCategorySelect(category)}>
-                  {category.title}
-                </DropdownItem>
-              ))}
-            />
-          </DataToolbarItem>
-          {filterCategories.map(category => (
-            <DataToolbarFilter
-              key={category.key}
-              chips={filterValues[category.key]}
-              deleteChip={onClearFilter}
-              categoryName={category.title}
-              showToolbarItem={currentFilterCategory.key === category.key}
-            >
-              <FilterControl category={category} value={filterValues[category.key]} />
-            </DataToolbarFilter>
-          ))}
-        </DataToolbarToggleGroup>
-      </DataToolbarContent>
-    </DataToolbar>
-  );
-
-  const { currentPageItems, paginationProps } = usePaginationState(rows, 10);
 
   ////////////////////////////////////////////////////////////////////////////////
 
-  // TODO: filters
+  const { currentPageItems, paginationProps } = usePaginationState(rows, 10);
+
   // TODO: sorting
 
   return (
@@ -322,7 +177,13 @@ const VolumesTable: React.FunctionComponent<IVolumesTableProps> = ({
       </GridItem>
       <GridItem>
         <Level>
-          <LevelItem>{filterToolbar}</LevelItem>
+          <LevelItem>
+            <FilterToolbar
+              filterCategories={filterCategories}
+              filterValues={filterValues}
+              setFilterValues={setFilterValues}
+            />
+          </LevelItem>
           <LevelItem>
             <Pagination widgetId="pv-table-pagination-top" {...paginationProps} />
           </LevelItem>
